Exercise02 changed int to double, almost perfect water splittin between flowers, but it shows bigger number than it should

+++ Week04/Day02/Exercise02/src/Plants.java
@@ -1,9 +1,9 @@
 public class Plants {
   String color;
-  int waterlevel;
-  int waterlimit;
+  double waterlevel;
+  double waterlimit;
 
-  public void watering(int waterIncome) {
+  public void watering(double waterIncome) {
     waterlevel = waterlevel + waterIncome;
   }
